DD-43 ♻ [Refactor] types 하위 폴더명 components로 변경
